requirements.txt
django
psycopg2
django-filter
Pillow
djangorestframework
